requirement.txt
streamlit
google-generativeai
python-dotenv
pandas
selenium
google-auth
beautifulsoup4